refact: Separação do Card com texto do Card com botões
