+++ src/styles/theme.ts
@@ -1,5 +1,6 @@
 import { createTheme } from '@mui/material/styles';
 
+
 export const theme = createTheme({
   palette: {
     mode: 'light',
